requirements.txt
torch
transformers>=4.38.0
accelerate
datasets
lm-eval
pyyaml
huggingface_hub
scipy
numpy
pandas
tqdm
llama-cpp-python  # For GGUF support
